feat(TTB-043): Refactored css and file structure
