+++ client/src/game/Square.jsx
