+++ client/src/game/game.css
@@ -1,0 +1,65 @@
+.board-row:after {
+    clear: both;
+    content: "";
+    display: table;
+}
+
+.status {
+    margin-bottom: 10px;
+}
+
+.square {
+    background: #fff;
+    border: 3px solid #222;
+    float: left;
+    font-size: 50px;
+    font-weight: bold;
+    line-height: 100px;
+    height: 100px;
+    margin-right: -3px;
+    margin-top: -3px;
+    padding: 0;
+    text-align: center;
+    width: 100px;
+}
+
+.container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    height: 100%;
+}
+
+.game {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+
+.game-info {
+    margin: 20px;
+    display: flex;
+    flex-direction: column;
+    align-content: center;
+    justify-content: center;
+    font-size: 20px;
+}
+
+.game-board {
+    display: flex;
+}
+
+.game-controls {
+    width: 100%;
+    display: flex;
+    padding-top: 35px;
+    justify-content: space-around;
+    flex-direction: row;
+    align-content: center
+}
+
+.game-control {
+    display: flex;
+}
