+++ client/src/index.css
@@ -1,117 +1,4 @@
 body {
     font: 14px "Century Gothic", Futura, sans-serif;
-    width: 100%;
-    height: 100%;
     margin: 0px;
-}
-
-ol,
-ul {
-    padding-left: 30px;
-}
-
-.board-row:after {
-    clear: both;
-    content: "";
-    display: table;
-}
-
-.status {
-    margin-bottom: 10px;
-}
-
-.square {
-    background: #fff;
-    border: 3px solid #222;
-    float: left;
-    font-size: 50px;
-    font-weight: bold;
-    line-height: 100px;
-    height: 100px;
-    margin-right: -3px;
-    margin-top: -3px;
-    padding: 0;
-    text-align: center;
-    width: 100px;
-}
-
-.square:focus {
-    outline: none;
-}
-
-.kbd-navigation .square:focus {
-    background: #ddd;
-}
-
-.container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-}
-
-.game {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-}
-
-.game-info {
-    margin: 20px;
-    display: flex;
-    flex-direction: column;
-    align-content: center;
-    justify-content: center;
-    font-size: 20px;
-}
-
-.game-board {
-    display: flex;
-}
-
-.game-controls {
-    width: 100%;
-    display: flex;
-    padding-top: 35px;
-    justify-content: space-around;
-    flex-direction: row;
-    align-content: center
-}
-
-.game-control {
-    display: flex;
-}
-
-.page-container {
-    display: flex;
-    min-height: 100vh;
-    height: 100%;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-}
-
-
-.header-container {
-    display: flex;
-    width: 100%;
-}
-
-.body-container {
-    display: flex;
-    flex-grow: 1;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-}
-
-.body-container-content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-around;
-    height: 100%;
-    padding-bottom: 50px;
 }
